# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BotAluguel Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans/active:
    get:
      operationId: getActivePlan
      tags: [plans]
      summary: Get user's active plan
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active plan or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivePlan"
  /plans/{planId}/activate:
    post:
      operationId: activatePlan
      tags: [plans]
      summary: Activate a plan using coins
      security:
        - bearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plan activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivePlan"
        "402":
          description: Insufficient coins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/pix:
    post:
      operationId: createPixCharge
      tags: [payments]
      summary: Create a PIX charge for coin top-up
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePixBody"
      responses:
        "201":
          description: PIX charge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PixCharge"
  /payments/pix/{txid}:
    get:
      operationId: checkPixStatus
      tags: [payments]
      summary: Check PIX payment status
      security:
        - bearerAuth: []
      parameters:
        - name: txid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PIX status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PixStatus"
  /payments/history:
    get:
      operationId: getPaymentHistory
      tags: [payments]
      summary: Get user's payment history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
...
  /admin/notifications:
    get:
      operationId: adminListNotifications
      tags: [admin]
      summary: List all notifications (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppNotification"
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppNotification"
  /notifications/unread-count:
    get:
      operationId: getUnreadCount
      tags: [notifications]
      summary: Get unread notification count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: number
                required:
                  - count
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required:
        - message
...
    ActivePlan:
      type: object
      properties:
        id:
          type: string
          nullable: true
        planId:
          type: string
          nullable: true
        planName:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        maxGroups:
          type: number
          nullable: true
        activatedAt:
          type: string
          nullable: true
          format: date-time
        expiresAt:
          type: string
          nullable: true
          format: date-time
      required: []
    CreatePixBody:
      type: object
      properties:
        amount:
          type: number
          description: Amount in BRL
      required:
        - amount
    PixCharge:
      type: object
      properties:
        txid:
          type: string
        amount:
          type: number
        coins:
          type: number
        status:
          type: string
        qrCodeBase64:
          type: string
          nullable: true
        copyPaste:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - txid
        - amount
        - coins
        - status
        - expiresAt
        - createdAt
    PixStatus:
      type: object
      properties:
        txid:
          type: string
        status:
          type: string
          enum: [pending, paid, expired, error]
        coins:
          type: number
          nullable: true
        paidAt:
          type: string
          nullable: true
          format: date-time
      required:
        - txid
        - status
    Payment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [topup, plan]
        amount:
          type: number
        coins:
          type: number
        status:
          type: string
          enum: [pending, paid, expired, error]
        txid:
          type: string
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          nullable: true
          format: date-time
      required:
        - id
        - userId
        - type
        - amount
        - coins
        - status
        - txid
        - createdAt
...
    AppNotification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - body
        - type
        - read
        - createdAt